Factor Sprite and DrawingSurface out of SongScore
These classes will be useful elsewhere and so have been moved into
ui/components.
